Backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL
# mysqlclient==2.2.0    # Uncomment for MySQL

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Image Processing
Pillow==10.1.0

# Firebase Cloud Messaging
firebase-admin==6.3.0

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
